requirements.txt
streamlit==1.31.0
openai==1.3.7
python-dotenv==1.0.0
httpx==0.24.1
setuptools==69.0.3
pip==24.0
wheel==0.42.0
google-genai==0.1.0
Pillow==10.2.0
